enh: enabled pointers to character(len=1) arrays
Now the storage of characters has been changed to
intrinsic character arrays.
This solves a couple of problems:

1. The requirement of iso_var_str has been made obsolete
2. A character(len=*) is stored as an array of character(len=1) (:)
3. One can retrieve any character array with a character(len=*) as though
   they were the same data-type.
   However, one can _only_ create a pointer to the character array
   as the character(len=*) pointer type REQUIRES specific length.

New tests has been added with this capability.
Note that to change between len=* and char arrays two functions has
been added, cpack and cunpack which packs a len=* to char array,
and the reverse routine, respectively.

Signed-off-by: Nick Papior <nickpapior@gmail.com>