maint: ABI breakage
Major change in the usage of the library.

Types are now renamed to variable_t and dictionary_t.
This makes a cleaner interface to external codes.

Also changed internal variables to have a suffix _ to make it
clear they are "private" or "on-the-fly" created.
This goes also for the *_inc files.
This should also make the build-system more robust.

Signed-off-by: Nick Papior <nickpapior@gmail.com>