Finalized cmake stuff and updated smeka build
Now everything should be compliant. Both the cmake
and the smeka builds should be compatible.

The only difference is that smeka does not provide cmake
targets for end-users.

A large restructure of cmake files was done.
Partially because I don't want this as a sub-project to
inflict problems on the hosting code. I will test
this momentarily with Siesta.

Fixed some bugs for smeka installations. It should now
produce the same tree-structure as CMake.

Still need to check whether the dist works, but that should
be done later.

Signed-off-by: Nick Papior <nickpapior@gmail.com>